refactor etherscan deployment script
