+++ package.json
@@ -68,7 +68,9 @@
     "deploy:stats": "npx hardhat run scripts/deploy.stats.ts",
     "deploy:vault": "npx hardhat run scripts/deploy.vault.ts",
     "deploy:pool:lido": "npx hardhat run scripts/deploy.pool.lido.ts",
-    "verify:pool:lido": "npx hardhat run scripts/verify.pool.lido.ts",
+    "verify:pool:lido": "cross-env TEMPUS_POOL_CONTRACT=LidoTempusPool npx hardhat run scripts/verify.pool.ts",
+    "verify:pool:yearn": "cross-env TEMPUS_POOL_CONTRACT=YearnTempusPool npx hardhat run scripts/verify.pool.ts",
+    "verify:pool:rari": "cross-env TEMPUS_POOL_CONTRACT=RariTempusPool npx hardhat run scripts/verify.pool.ts",
     "increase-time": "npx hardhat run scripts/increaseTime.local.ts --network localhost",
     "mythx-install": "pip install mythx-cli",
     "mythx-amm": "mythx --ci --yes analyze --wait --solc-version 0.7.6 contracts/amm/TempusAMM.sol contracts/amm/TempusAMMUserDataHelpers.sol",
